implemented basic permission based user authentication system

+++ src/main/java/com/surrey/com3014/group5/controllers/UserController.java
@@ -1,5 +1,6 @@
 package com.surrey.com3014.group5.controllers;
 
+import com.surrey.com3014.group5.exceptions.NotFoundException;
 import com.surrey.com3014.group5.models.impl.User;
 import com.surrey.com3014.group5.services.user.UserService;
 import org.slf4j.Logger;
@@ -8,8 +9,8 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.*;
 
-import javax.persistence.EntityNotFoundException;
 import javax.servlet.http.HttpServletResponse;
+import java.util.Optional;
 
 /**
  * Spring MVC controller to handle user registration and management.
@@ -55,9 +56,14 @@ public class UserController {
     @ResponseStatus(value = HttpStatus.NO_CONTENT)
     @ResponseBody
     public void deleteById(@PathVariable("id") long id, HttpServletResponse response) {
-        User user = userService.findOne(id);
-        response.setHeader("Location", "/users");
-        userService.delete(user);
+        Optional<User> userOptional = userService.findOne(id);
+        if (userOptional.isPresent()) {
+            User user = userOptional.get();
+            response.setHeader("Location", "/users");
+            userService.delete(user);
+        }
+
+        throw new NotFoundException("The requested user does not exist");
     }
 
     /**
@@ -67,11 +73,11 @@ public class UserController {
     @ResponseStatus(value = HttpStatus.NO_CONTENT)
     @ResponseBody
     public void deleteByUsername(String username) {
-        User user = userService.findByUsername(username);
-        if (user == null) {
-            throw new EntityNotFoundException("Entity with username: " + username + " not found");
+        Optional<User> user = userService.findByUsername(username);
+        if (user.isPresent()) {
+            userService.delete(user.get());
         }
-        userService.delete(user);
+        throw new NotFoundException("User with username: " + username + " does not exist");
     }
 
     /**
@@ -81,12 +87,15 @@ public class UserController {
     @ResponseStatus(value = HttpStatus.NO_CONTENT)
     @ResponseBody
     public void updateUser(@PathVariable("id") long id, String password, String email, String name) {
-        User user = userService.findOne(id);
-        //String encodedPassword = getPasswordEncoder().encode(password);
-        user.setPassword(password);
-        user.setEmail(email);
-        user.setName(name);
-        userService.update(user);
+        Optional<User> userOptional = userService.findOne(id);
+        if (userOptional.isPresent()) {
+            User user = userOptional.get();
+            user.setPassword(password);
+            user.setEmail(email);
+            user.setName(name);
+            userService.update(user);
+        }
+        throw new NotFoundException("The requested user does not exist");
     }
 
     /**
@@ -96,11 +105,11 @@ public class UserController {
     @ResponseStatus(value = HttpStatus.OK)
     @ResponseBody
     public User getByUsername(String username) {
-        User user = userService.findByUsername(username);
-        if (user == null) {
-            throw new EntityNotFoundException("Entity with username: " + username + " not found");
+        Optional<User> user = userService.findByUsername(username);
+        if (user.isPresent()) {
+            return user.get();
         }
-        return user;
+        throw new NotFoundException("The requested user with username: " + username + " does not exist");
     }
 
     /**
@@ -110,11 +119,11 @@ public class UserController {
     @ResponseStatus(value = HttpStatus.OK)
     @ResponseBody
     public User getByEmail(String email) {
-        User user = userService.findByEmail(email);
-        if (user == null) {
-            throw new EntityNotFoundException("Entity with email: " + email + "not found");
+        Optional<User> user = userService.findByEmail(email);
+        if (user.isPresent()) {
+            return user.get();
         }
-        return user;
+        throw new NotFoundException("The requested user with email: " + email + " does not exist");
     }
 
     public UserService getUserService() {
